# backend/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Fight Club API",
    "description": "Complete API documentation for the AI Fight Club application - a platform for AI coding challenges and evaluation",
    "version": "1.0.0",
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    },
    "contact": {
      "name": "Support",
      "url": "https://aifightclub.example.com",
      "email": "support@aifightclub.example.com"
    },
    "termsOfService": "https://aifightclub.example.com/terms",
    "x-api-versioning": {
      "strategy": "uri-path",
      "current": "v1",
      "deprecated": [],
      "sunset": []
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current server (relative path)"
    }
  ],
...
  "paths": {
...
    "/adaptive/recommendations": {
      "get": {
        "summary": "Get adaptive recommendations",
        "description": "Retrieves personalized adaptive learning recommendations for the current user",
        "operationId": "getAdaptiveRecommendations",
        "tags": [
          "Adaptive"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by recommendation type",
            "schema": {
              "type": "string",
              "enum": [
                "challenge",
                "resource",
                "skill_focus",
                "difficulty_adjustment"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by recommendation status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "dismissed",
                "completed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recommendations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdaptiveRecommendation"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationParams"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/adaptive/dynamic-challenge": {
      "post": {
        "summary": "Generate dynamic challenge",
        "description": "Generates a challenge dynamically based on user's skill level and learning path",
        "operationId": "generateDynamicChallenge",
        "tags": [
          "Adaptive",
          "Challenges"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "focusArea": {
                    "type": "string",
                    "description": "Specific area to focus on (optional)"
                  },
                  "preferredDifficulty": {
                    "type": "string",
                    "enum": [
                      "beginner",
                      "intermediate",
                      "advanced",
                      "expert"
                    ],
                    "description": "User's preferred difficulty level (optional)"
                  },
                  "includeConcepts": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Specific concepts to include (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dynamic challenge created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Challenge"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication. Use the token from the login response. Tokens expire after 24 hours."
      }
    },
    "parameters": {
      "limitParam": {
        "name": "limit",
        "in": "query",
        "description": "Maximum number of items to return per page (default: 20, max: 100)",
        "schema": {
          "type": "integer",
          "default": 20,
          "minimum": 1,
          "maximum": 100
        }
      },
      "offsetParam": {
        "name": "offset",
        "in": "query",
        "description": "Number of items to skip (for pagination)",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized - Invalid or missing token"
                }
              }
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "The request data failed validation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Validation error"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "PaginationParams": {
        "type": "object",
        "description": "Parameters for pagination",
        "properties": {
          "limit": {
            "type": "integer",
            "default": 20,
            "minimum": 1,
            "maximum": 100,
            "description": "Maximum number of items to return"
          },
          "offset": {
            "type": "integer",
            "default": 0,
            "minimum": 0,
            "description": "Number of items to skip"
          },
          "sort": {
            "type": "string",
            "description": "Field to sort by (prefix with - for descending order)"
          }
        }
      },
...
      "Challenge": {
        "type": "object",
        "description": "Challenge information",
        "required": [
          "id",
          "content",
          "userEmail",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the challenge"
          },
          "content": {
            "type": "string",
            "description": "The challenge content/question"
          },
          "userEmail": {
            "type": "string",
            "format": "email",
            "description": "Email of the user the challenge is for"
          },
          "focusArea": {
            "type": "string",
            "description": "The focus area of the challenge"
          },
          "challengeType": {
            "type": "string",
            "description": "The type of challenge"
          },
          "formatType": {
            "type": "string",
            "description": "The format of the challenge"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced",
              "expert"
            ],
            "description": "Difficulty level of the challenge"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "submitted",
              "evaluated"
            ],
            "description": "Current status of the challenge"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the challenge was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the challenge was last updated"
          }
        }
      },
...
      "AdaptiveRecommendation": {
        "type": "object",
        "description": "Adaptive learning recommendation for a user",
        "required": [
          "id",
          "userId",
          "type",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the recommendation"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user the recommendation is for"
          },
          "type": {
            "type": "string",
            "enum": [
              "challenge",
              "resource",
              "skill_focus",
              "difficulty_adjustment"
            ],
            "description": "Type of recommendation"
          },
          "contentId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the recommended content (challenge, resource, etc.)"
          },
          "reason": {
            "type": "string",
            "description": "Explanation for why this recommendation was made"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10,
            "description": "Priority of the recommendation (1-10)"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "dismissed",
              "completed"
            ],
            "description": "Current status of the recommendation"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata about the recommendation",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the recommendation was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the recommendation was last updated"
          }
        }
      },
...
    }
  }
}